Tweaks mostly involving size_t warnings.
git-svn-id: file:///mnt/svn/rtmath@176 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath/atmos.cpp
@@ -98,7 +98,7 @@ namespace rtmath {
 			gasnames.push_back(buffer);
 		}
 #ifdef _WIN32
-		numgases = gasnames.size(); 
+		numgases = (unsigned int) gasnames.size(); 
 #else // parser.good() repeats the last value. Aarg!
 		numgases = gasnames.size() - 1;
 #endif
